In Progress

+++ TestSutraUsers/src/main/java/com/tathvatech/survey/oid/ResponseSubmissionBookmarkOID.java
@@ -2,10 +2,11 @@ package com.tathvatech.survey.oid;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.tathvatech.ts.core.common.EntityTypeEnum;
-import com.tathvatech.ts.core.common.OID;
+import com.tathvatech.common.enums.EntityTypeEnum;
+import com.tathvatech.user.OID.OID;
 
-public class ResponseSubmissionBookmarkOID extends OID{
+
+public class ResponseSubmissionBookmarkOID extends OID {
 
 	public ResponseSubmissionBookmarkOID()
 	{
@@ -45,7 +46,7 @@ public class ResponseSubmissionBookmarkOID extends OID{
 	
 	@Override
 	public int hashCode() {
-		return super.getPk();
+		return (int) super.getPk();
 	}
 
 }
