+++ TestSutraUsers/src/main/java/com/tathvatech/timetracker/oid/ShiftInstanceOID.java
@@ -2,10 +2,11 @@ package com.tathvatech.timetracker.oid;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.tathvatech.ts.core.common.EntityTypeEnum;
-import com.tathvatech.ts.core.common.OID;
+import com.tathvatech.common.enums.EntityTypeEnum;
+import com.tathvatech.user.OID.OID;
 
-public class ShiftInstanceOID extends OID{
+
+public class ShiftInstanceOID extends OID {
 
 	@JsonCreator
 	public ShiftInstanceOID(@JsonProperty("pk") int pk)
@@ -40,7 +41,7 @@ public class ShiftInstanceOID extends OID{
 	
 	@Override
 	public int hashCode() {
-		return super.getPk();
+		return (int) super.getPk();
 	}
 
 }
